Refactor InboxPage out of AdaptiveNav
